processed thijs review comments
