+++ api/app/signals/apps/signals/factories/__init__.py
@@ -10,6 +10,7 @@ from .priority import PriorityFactory
 from .question import QuestionFactory
 from .reporter import ReporterFactory
 from .signal import SignalFactory, SignalFactoryValidLocation, SignalFactoryWithImage
+from .signal_departments import SignalDepartmentsFactory
 from .source import SourceFactory
 from .status import StatusFactory
 from .status_message_template import StatusMessageTemplateFactory
@@ -35,6 +36,7 @@ __all__ = [
     'SignalFactory',
     'SignalFactoryValidLocation',
     'SignalFactoryWithImage',
+    'SignalDepartmentsFactory',
     'SourceFactory',
     'StatusFactory',
     'StatusMessageTemplateFactory',
