+++ api/app/signals/apps/signals/models/routing_expression.py
@@ -31,4 +31,4 @@ class RoutingExpression(models.Model):
     order = models.PositiveIntegerField(default=0, db_index=True)
     is_active = models.BooleanField(default=False)
 
-    objects = RoutingExpressionManager()
+    actions = RoutingExpressionManager()
