+++ api/app/signals/apps/signals/querysets.py
@@ -9,6 +9,6 @@ class SignalQuerySet(QuerySet):
     def filter_for_user(self, user):
         if not user.is_superuser and not user.has_perm('signals.sia_can_view_all_categories'):
             # We are not a superuser and we do not have the "show all categories" permission
-            return self.filter(self.permission_service.make_permisson_condition(user))
+            return self.filter(self.permission_service.make_permisson_condition_for_user(user))
 
         return self.all()
